requirements.txt
streamlit==1.41.1
langchain==0.3.9
langchain-openai==0.2.11
python-dotenv==1.0.1
openai==1.57.2